Subcontractor Auto Sync and Reminder conversation thread

+++ Repositories/Interfaces/ICommon.cs
@@ -9,7 +9,7 @@ namespace UnibouwAPI.Repositories.Interfaces
 
         Task<IEnumerable<Person>> GetAllPerson();
         Task<Person?> GetPersonById(Guid id);
-        Task<int> CreatePerson(Person person);
+        Task<long> CreatePerson(Person person);
 
         Task<IEnumerable<SubcontractorWorkItemMapping>> GetAllSubcontractorWorkItemMapping(bool onlyActive = false);
         Task<List<SubcontractorWorkItemMapping?>> GetSubcontractorWorkItemMappingById(Guid id);
